cilium: ipsec, insert 'ip rules' for ipsec
IPSec will be isolated from the "normal" routing tables via a set of
marks to direct traffic to an IPSec table. When an endpoint has IPSec
enabled the datapath will first set the mark bit to endicate
MARK_ENCRYPTED or MARK_DECRYPTED to depending on expected direction.
Once the encryption or decryption has occured the mark fields will
be cleared and normal processing will happen.

In this patch add the mark based 'ip rules' to segregate the IPSec
traffic.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>