kvstore: Release local kvstore lock after timeout
Cilium does distributed locking via etcd. For this purpose we first acquire a
local lock indexed by the etcd key path. This is needed because etcd locks only
protect against other clients and not from mutual access from the same etcd
client. We then acquire the lock in etcd itself. We store locally that we are
currently holding the etcd lock. When the etcd lock is unlocked, this local
state is removed.

For some reason, it’s unclear why, the local state was left behind and the
local unlocking never took place. etcd was identified to have crashed several
times based on logs which could have caused the etcd client returning something
unexpected.

When unable to acquire the local lock, we ignore this state and ask etcd
directly. Unfortunately the original request requiring the etcd lock, in this
case identity allocation, has typically already timed out. The current code
never corrects the local state until the cilium pod is restarted.

Fixes: #6667

Signed-off-by: Thomas Graf <thomas@cilium.io>